Rename PagesController to FinishController
Since it contains only the two possible finish routes, eligible or
ineligible. This way it acts as the end of the flow, in the same way
that the StartController acts as the beginning.
