+++ app/views/eligibility_interface/finish/eligible.html.erb
